refactor: split the REF grammar into field-ref-parse to fit the line budget
